improve output readability

+++ Mesh.cpp
@@ -58,7 +58,7 @@ void Mesh::readSnapshot(const std::string& snapshot) {
         std::vector<float> cellVelocities1D(dims[0] * dims[1]);
         velocitiesDataset.read(cellVelocities1D.data(), H5::PredType::NATIVE_FLOAT);
 
-        cellVelocities.resize(dims[0], std::vector<double>(dims[1]));
+        cellVelocities.resize(dims[0], std::vector<float>(dims[1]));
 
         for (size_t row = 0; row < dims[0]; ++row) {
              for (size_t col = 0; col < dims[1]; ++col) {
